% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candeia.R
\docType{data}
\name{candeia}
\alias{candeia}
\title{candeia dataset example}
\description{
candeia dataset example
}
\details{
candeia
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
