% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunc.R
\name{contr.tuk}
\alias{contr.tuk}
\title{Tukey's contrast matrix}
\usage{
contr.tuk(x)
}
\arguments{
\item{x}{a vector of means.}
}
\value{
The matrix of contrasts with treatments in row names is returned
}
\description{
Compute Tukey's contrast matrix
}
\details{
Computes the matrix of contrasts for comparisons of mean levels.
}
\examples{
x <- c(10, 5, 8, 4, 12, 18)
contr.tuk(x)

}
