% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crd_simulated.R
\docType{data}
\name{crd_simulated}
\alias{crd_simulated}
\title{Simulated data set for CRD}
\description{
Simulated data set for CRD
}
\details{
This dataset was simulated under a gaussian random field containing 15 treatments and 8 replication with mean 10.
}
\author{
Lucas Castro \email{lrcastro@estudante.ufla.br}
}
