% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbd_simulated.R
\docType{data}
\name{rcbd_simulated}
\alias{rcbd_simulated}
\title{Simulated data set for RCBD}
\description{
Simulated data set for RCBD
}
\details{
This dataset was simulated under a gaussian random field containing 15 treatments and 3 blocks with 3 replication.
}
\author{
Lucas Castro \email{lrcastro@estudante.ufla.br}
}
