% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_aov.R
\name{sar_aov}
\alias{sar_aov}
\title{Using a SAR model to handle spatial dependence in a aov model.}
\usage{
sar_aov(formula, coord, data, seq.radius = NULL)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{coord}{A matrix or data.frame of point coordinates or a SpatialPoints Object.}

\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{seq.radius}{A complex vector containing a radii sequence used to set the neighborhood pattern.
The default sequence has ten numbers from 0 to half of the maximum distance between the samples.}
}
\value{
\code{sar_aov} returns an object of \code{\link[base]{class}} "SARaov".
The functions summary and anova are used to obtain and print a summary and analysis of variance
table of the results.
An object of class "SARaov" is a list containing the following components:

\item{DF}{degrees of freedom of rho, treatments, residual and total.}
\item{SS}{sum of squares of rho, treatments and residual.}
\item{rho}{the autoregressive parameter.}
\item{Par}{data.frame with the radius tested and its AIC.}
\item{modelAdj}{model of class \code{\link[stats]{aov}} using the adjusted response.}
\item{modelstd}{data frame containing the ANOVA table using non-adjusted response.}
\item{namey}{response variable name.}
}
\description{
Fit a completely randomized design when the experimental units have some degree of
spatial dependence using a Spatial Lag Model (SAR).
}
\examples{
\dontrun{
data("carrancas")
resp <- carrancas$DAP16
treat <- carrancas$T
coord <- cbind(carrancas$X, carrancas$Y)
mdl <- aov(resp ~ treat, data = carrancas)
cv<-sar_aov(mdl, coord)
cv

#Summary for class SARanova
summary(cv)

#Anova for class SARanova
anova(cv)
}

}
\references{
Long, D. S. "Spatial statistics for analysis of variance of agronomic field trials."
Practical handbook of spatial statistics. CRC Press, Boca Raton, FL (1996): 251-278.
}
