% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_rcbd.R
\name{sar_rcbd}
\alias{sar_rcbd}
\title{Using a SAR model to handle spatial dependence in a Randomized Complete Block Design}
\usage{
sar_rcbd(resp, treat, coord, seq.radius = NULL)
}
\arguments{
\item{resp}{Numeric or complex vector containing the values of response variable.}

\item{treat}{Numeric or complex vector containing the treatment applied to each experimental unit.}

\item{block}{Numeric or complex vector specifying the blocks.}

\item{coord}{Matrix of point coordinates or a SpatialPoints Object.}

\item{seq.radius}{Complex vector containing a radii sequence used to set the neighborhood pattern.
The default sequence has ten numbers from 0 to half of the maximum distance between the samples.}
}
\value{
\code{sar_rcbd} returns an object of \code{\link[base]{class}} "SARanova".
The functions summary and anova are used to obtain and print a summary and analysis of variance
table of the results.
An object of class "SARanova" is a list containing the following components:

\item{DF}{degrees of freedom of rho, treatments, residual and total.}
\item{SS}{sum of squares of rho, treatments and residual.}
\item{Fc}{F statistic calculated for treatment.}
\item{p.value}{p-value associated to F statistic for treatment.}
\item{rho}{the autoregressive parameter.}
\item{Par}{data.frame with the radii tested and its AIC.}
\item{y_orig}{vector of response.}
\item{treat}{vector of treatment applied to each experimental unit.}
\item{modelAdj}{model of class \code{\link[stats]{aov}} using the adjusted response.}
\item{namey}{response variable name.}
\item{namex}{treatment variable name.}
\item{modelstd}{data frame containing the ANOVA table using non-adjusted response.}
}
\description{
Fit a randomized complete block design when the experimental units have some degree of
spatial dependence using a Spatial Lag Model (SAR).
}
\examples{
\dontrun{
data("carrancas")
resp <- carrancas$DAP16
treat <- carrancas$T
block <- carrancas$Bloco
coord <- cbind(carrancas$X, carrancas$Y)
cv<-sar_rcbd(resp, treat, block, coord)
cv

#Summary for class SARanova
summary(cv)

#Anova for class SARanova
anova(cv)

#Test based on multivariate t-student distribution
spMVT(cv)

#Tukey's test
spTukey(cv)

#Scott-Knott test
spScottKnott(cv)

}

}
\references{
Scolforo, Henrique Ferraço, et al. "Autoregressive spatial analysis and individual
tree modeling as strategies for the management of Eremanthus erythropappus." Journal of
forestry research 27.3 (2016): 595-603.
}
