% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCrossvalid.R
\name{spCrossvalid}
\alias{spCrossvalid}
\title{Cross-validation by kriging}
\usage{
spCrossvalid(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link[spANOVA]{spVariofit}}.}

\item{...}{further arguments to be passed to \code{\link[geoR]{xvalid}} function.}
}
\description{
Compute cross-validation for an object of class \code{\link[spANOVA]{spVariofit}}.
}
\details{
This function is a wrapper to \code{\link[geoR]{xvalid}} function of the package geoR.
Please check its documentation for additional information.
}
