% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTukey.R
\name{spTukey}
\alias{spTukey}
\alias{spTukey.SARcrd}
\alias{spTukey.SARrcbd}
\alias{spTukey.GEOanova}
\title{Compute Tukey Honest Significant Differences for a Spatially Correlated Model}
\usage{
spTukey(x, sig.level = 0.05)

\method{spTukey}{SARcrd}(x, sig.level = 0.05)

\method{spTukey}{SARrcbd}(x, sig.level = 0.05)

\method{spTukey}{GEOanova}(x, sig.level = 0.05)
}
\arguments{
\item{x}{A fitted model object of class SARcrd, SARrcbd or GEOanova.}

\item{sig.level}{A numeric value between zero and one giving the significance
level to use.}
}
\value{
a data frame containing the mean and its group.
}
\description{
Perform multiple comparisons of means treatments based on the Studentized range
statistic when the errors are spatially correlated.
}
\details{
For objects of class SARcrd or SARrcbd this function performs the standard
Tukey's ‘Honest Significant Difference’ method provided by the function
\code{\link[stats]{TukeyHSD}} on the adjusted response.

For objects of class GEOanova, the method is modified to take into account
the spatial dependence among the observations. First, we estimate a
contrast matrix (\eqn{C}) using cont.tuk function and then after estimate the
spatial mean of each treatment (\eqn{\mu_i}) we can assess the significance of
the contrast by

\deqn{|c_i \mu_i| > {HSD}_i}

where \eqn{HSD_i = q(\alpha, k, \nu) * sqrt(0.5*{w}_ii)} and
\eqn{k} is the number of treatments, \eqn{\alpha} is the level of significance,
\eqn{\nu} is the degree of freedom of the model, \eqn{{w}_ii} is the variance of
the i-th contrast.
}
\references{
NOGUEIRA, C. H. Testes para comparações múltiplas de
médias em experimentos com tendência e dependência espacial.
142 f. Tese (Doutorado em Estatística e Experimentação
Agropecuária) | Universidade Federal de Lavras, Lavras, 2017
}
