% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crd_simulated.R
\docType{data}
\name{crd_simulated}
\alias{crd_simulated}
\title{crd_simulated dataset example}
\description{
crd_simulated dataset example
}
\details{
crd_simulated
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
