% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spANOVAapp.R
\name{spANOVAapp}
\alias{spANOVAapp}
\title{Shiny app for spANOVA}
\usage{
spANOVAapp(external = TRUE)
}
\arguments{
\item{external}{logical. If true, the system's default web browser will be
launched automatically after the app is started.}
}
\description{
Shiny app for analysis of variance with spatially correlated errors
}
